move extractSourceHeaders to more appropriate place
